Moving ViewEditUser functions into ViewEditUsers
